+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Now patterns are also expanded in the out_dir name.
 - Global options to control the date format.
 - Outputs can use the options of other outputs as base (extend them). (#112)
+- Another experimental mechanism to change 3D models according to the variant.
+  (#103)
 
 ### Changed
 - Internal BoM: now components with different Tolerance, Voltage, Current